requirements.txt
# AI Interview Assistant - Requirements
# FastAPI Backend with AI/ML Features for Interview Analysis
# Windows Compatible Versions

# Core Web Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database & ORM
SQLAlchemy==2.0.23
greenlet==3.2.4

# Data Validation & Serialization
pydantic==2.5.0
pydantic-core==2.14.1
email-validator==2.1.0

# File Upload & Processing
python-multipart==0.0.6
aiofiles==23.2.1

# AI/ML Libraries - Windows Compatible Versions
# Core ML Libraries
torch==2.1.1
torchaudio==2.1.1
transformers==4.36.0
tokenizers==0.15.2
safetensors==0.6.2
huggingface-hub==0.35.1

# Whisper for Audio Transcription
openai-whisper==20231117
numba==0.58.1
llvmlite==0.41.1

# Natural Language Processing
nltk==3.8.1
scikit-learn==1.3.2
scipy==1.15.3
joblib==1.5.2
threadpoolctl==3.6.0

# Data Processing & Analysis
pandas==2.1.4
matplotlib==3.8.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0

# NumPy - Compatible with TensorFlow and other ML libraries
numpy>=1.26.0

# Mathematical Libraries
sympy==1.14.0
mpmath==1.3.0
networkx==3.4.2

# PDF Generation
reportlab==4.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==46.0.1
cffi==2.0.0
pycparser==2.23

# HTTP Client & Networking
requests==2.31.0
httpx==0.25.2
httpcore==1.0.9
h11==0.16.0
httptools==0.6.4
websockets==15.0.1
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Environment & Configuration
python-dotenv==1.0.0

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2

# Utilities
click==8.3.0
colorama==0.4.6
packaging==25.0
tqdm==4.67.1
more-itertools==10.8.0
regex==2025.9.18
tiktoken==0.11.0
filelock==3.19.1
fsspec==2025.9.0
typing-extensions==4.15.0
annotated-types==0.7.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
iniconfig==2.1.0
pluggy==1.6.0
tomli==2.2.1
exceptiongroup==1.3.0
anyio==3.7.1
sniffio==1.3.1

# File Watching for Development
watchfiles==1.1.0

# Additional Security
pyasn1==0.6.1
rsa==4.9.1
dnspython==2.8.0
ecdsa==0.19.1

# YAML Support
PyYAML==6.0.2

# Setuptools
setuptools==65.5.0